requirements.txt
requests==2.18.4
setuptools==38.4.0
six==1.11.0
websocket_client==0.55.0
